smartpharmacist/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
channels==4.0.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.3
django-restframework==0.0.1
djangorestframework==3.15.2
fonttools==4.50.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
django-browser-reload==1.13.0 
django-tailwind==3.8.0
